+++ cocos2d/core/renderer/webgl/assemblers/assembler.js
