+++ cocos2d/core/renderer/webgl/assemblers/graphics/earcut.js
