+++ cocos2d/core/renderer/webgl/assemblers/index.js
