+++ cocos2d/core/renderer/webgl/assemblers/sprite/simple.js
@@ -23,8 +23,8 @@
  THE SOFTWARE.
  ****************************************************************************/
 
-const js = require('../../../platform/js');
-const dynamicAtlasManager = require('../../utils/dynamic-atlas/manager');
+const js = require('../../../../platform/js');
+const dynamicAtlasManager = require('../../../utils/dynamic-atlas/manager');
 
 module.exports = {
     useModel: false,
