+++ cocos2d/core/renderer/webgl/assemblers/sprite/sliced.js
@@ -23,8 +23,8 @@
  THE SOFTWARE.
  ****************************************************************************/
 
-const Sprite = require('../../../components/CCSprite');
-const dynamicAtlasManager = require('../../utils/dynamic-atlas/manager');
+const Sprite = require('../../../../components/CCSprite');
+const dynamicAtlasManager = require('../../../utils/dynamic-atlas/manager');
 const FillType = Sprite.FillType;
 
 module.exports = {
